requirements.txt
streamlit
pandas
datetime
alpaca-py
altair
plotly
matplotlib
seaborn
numpy